Linux-PAM: Autospec creation for version 1.3.1
The files present in this folder are all completely commented out templates.

I'm moving them to the -doc package so there's an example available for users who
want or need to reference them. None of these files is related to modules that
are used by default, these are all optional and non-default modules.